Reacts Projects/trw-mvp/server/server/main.py
/*
 Copyright (c) 2026 Ashraf Morningstar
 These are personal recreations of existing projects, developed by Ashraf Morningstar
 for learning and skill development.
 Original project concepts remain the intellectual property of their respective creators.
 Repository: https://github.com/AshrafMorningstar
*/

#
# -----------------------------------------------------------------------------
# @author      Ashraf Morningstar
# @github      https://github.com/AshrafMorningstar
# @repository  Project Graveyard - The Ultimate Archive
# @quote       "Code that defines the future. Designed to inspire."
# -----------------------------------------------------------------------------
#

from fastapi import FastAPI, UploadFile, File, HTTPException, Path
from fastapi.responses import JSONResponse
from pydantic import BaseModel
import uuid, yaml

app = FastAPI(title="Titan Reality Weaver API (Stub)", version="1.0.0", openapi_url="/v1/openapi.json", docs_url="/v1/docs", redoc_url="/v1/redoc")

class ConnectorSpec(BaseModel):
    name: str
    type: str
    config: dict

class ConnectorResponse(BaseModel):
    id: str
    status: str

class ModelMeta(BaseModel):
    id: str
    name: str = None
    version: str = None
    runtime: str = None
    created_at: str = None
    signature: str = None

class TwinSummary(BaseModel):
    twin_id: str
    status: str

class RunSpec(BaseModel):
    mode: str = "simulate"
    start_time: str = None
    duration_seconds: int = 60
    parameters: dict = {}

class RunResponse(BaseModel):
    run_id: str
    status: str

@app.post("/v1/connectors", response_model=ConnectorResponse, status_code=201)
async def register_connector(spec: ConnectorSpec):
    # In real system: validate connector and persist
    return ConnectorResponse(id=str(uuid.uuid4()), status="registered")

@app.post("/v1/models", response_model=ModelMeta, status_code=201)
async def upload_model(manifest: str = None, file: UploadFile = File(None)):
    # Prototype: accept upload, store minimal metadata
    mid = str(uuid.uuid4())
    name = file.filename if file else "inline-manifest"
    return ModelMeta(id=mid, name=name, version="0.0.1", runtime="wasm", created_at="2025-01-01T00:00:00Z", signature="stub-signature")

@app.get("/v1/models")
async def list_models():
    return [ {"id":"model-1","name":"motor_surrogate","version":"1.0.0","runtime":"wasm"} ]

@app.get("/v1/models/{model_id}")
async def get_model(model_id: str = Path(...)):
    return {"id": model_id, "name":"example","version":"1.0.0","runtime":"wasm"}

@app.post("/v1/twins", response_model=TwinSummary, status_code=201)
async def create_twin(body: bytes = File(...)):
    # body will contain YAML twin-compose; validate basic structure
    try:
        spec = yaml.safe_load(body)
        twin_id = spec.get("twin_id", str(uuid.uuid4()))
        return TwinSummary(twin_id=twin_id, status="created")
    except Exception as e:
        raise HTTPException(status_code=400, detail=str(e))

@app.post("/v1/twins/{twin_id}/run", response_model=RunResponse, status_code=202)
async def start_run(twin_id: str, run_spec: RunSpec):
    # In prototype: return run id and accepted status
    return RunResponse(run_id=str(uuid.uuid4()), status="accepted")

@app.get("/v1/runs/{run_id}/logs")
async def run_logs(run_id: str, cursor: str = None):
    # Return sample trace page
    return {
        "run_id": run_id,
        "items": [
            {"timestamp":"2025-01-01T00:00:00Z","type":"info","payload":{"msg":"run started"}},
            {"timestamp":"2025-01-01T00:00:05Z","type":"metric","payload":{"throughput":120}}
        ],
        "next_cursor": None
    }
