Reacts Projects/trw-mvp/server/requirements.txt
/*
 Copyright (c) 2026 Ashraf Morningstar
 These are personal recreations of existing projects, developed by Ashraf Morningstar
 for learning and skill development.
 Original project concepts remain the intellectual property of their respective creators.
 Repository: https://github.com/AshrafMorningstar
*/

fastapi
uvicorn[standard]
pyyaml
pydantic
